Keep track of paperd protocol file in Paperclip as well
paperd reads this file to determine the protocol version the jar
supports. Since Paperclip wraps our output jar too we need to include
this protocol version file in Paperclip as well.

Also this commits refactors and modernizes the project in general,
updating / removing dependencies, cleaning up the code, and moving the
code to the io.papermc package.